vdpa/sfc: add support for mcdi IOVA remap
vDPA driver allocates a MCDI buffer and maps it for DMA in the IO address
space of the virtual function, and the virtio-net driver independently
allocates virtqueues and uses their corresponding guest IOVA(s). There is
no guarantee that the values for these two addresses will be unique and
vDPA might need to relocate mcdi IOVA.

To resolves the problem of overlap between IOVA(s) allocated by vDPA and
qemu, three changes have been made to the vDPA driver.

1) Cache all known IOVA(s) in a linked list and add functions
   to check overlap and resolve IOVA opverlaps.
2) Check for overlap
3) If overlap is found, find a new IOVA and invoke efx_mcdi_dma_remap

Signed-off-by: Abhimanyu Saini <absaini@amd.com>