net/nfp: support reading linkspeed from port table
The NFP PMD driver retrieves the linkspeed by reading the
NFP_NET_CFG_STS_LINK register. This register is not set by all versions
of the firmware.

Add a second preferred method to read the linkspeed from the port table
instead, while keeping the old lookup method as a fallback in case that
is not supported.

Signed-off-by: Long Wu <long.wu@corigine.com>
Reviewed-by: Chaoyong He <chaoyong.he@corigine.com>
Reviewed-by: Niklas Söderlund <niklas.soderlund@corigine.com>